% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_clust.R
\name{em_clust}
\alias{em_clust}
\title{IMplements EM algorithm for clustering data based on multivariate normal distributions}
\usage{
em_clust(dat, k)
}
\arguments{
\item{dat}{dataframe to perform clustering on}

\item{k}{number of clusters}
}
\value{
A list object with the p, probabilities of the ith datapoint belonging to the jth group,
mu, a dataframe of the centers of the k multivariate normal sources, and sigma, the covariance matrices
of each of the mvn sources
}
\description{
IMplements EM algorithm for clustering data based on multivariate normal distributions
}
