% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{getminfromdistmat}
\alias{getminfromdistmat}
\title{Helper function to get row and column from dist matrix}
\usage{
getminfromdistmat(distmat, nrowx)
}
\arguments{
\item{nrowx}{number of rows in x}

\item{distmax}{dataframe to perform clustering on}
}
\value{
loc of min
}
\description{
Helper function to get row and column from dist matrix
}
