% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{hier_clust}
\alias{hier_clust}
\title{Implements hierarchical clustering}
\usage{
hier_clust(dat, k)
}
\arguments{
\item{dat}{dataframe to perform clustering on}

\item{k}{number of clusters}
}
\value{
list of size k of strings with groups in them lol
}
\description{
Implements hierarchical clustering
}
