% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{Implements simple linear regression by hand}
\usage{
k_means(x, k)
}
\arguments{
\item{x}{A data frame}

\item{k}{sample size for initial clustersnumber of iterations}

\item{iters}{number of iterations}
}
\value{
A data frame
}
\description{
Implements simple linear regression by hand
}
