% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{matrix_find_min}
\alias{matrix_find_min}
\title{Implements simple linear regression by hand}
\usage{
matrix_find_min(matr)
}
\arguments{
\item{dataset}{A data frame}
}
\value{
clusters
}
\description{
Implements simple linear regression by hand
}
