% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_means.R
\name{k_means}
\alias{k_means}
\title{Implements naive k means algorithm (optionally for first two principle componenets of a dataset)}
\usage{
k_means(dat, nk, pca2 = F)
}
\arguments{
\item{dat}{dataframe to perform clustering on}

\item{nk}{number of clusters}

\item{pca2}{boolean whether or not to use pca}
}
\value{
A list object with groups by row, ss within, ss between, sstotal
}
\description{
Implements naive k means algorithm (optionally for first two principle componenets of a dataset)
}
